fix editor fallback across launch paths
