+++ apps/zerocode/src/editor.rs
@@ -1,0 +1,32 @@
+pub(crate) fn editor_from_env_or_path() -> Option<String> {
+    std::env::var("VISUAL")
+        .ok()
+        .filter(|value| !value.trim().is_empty())
+        .or_else(|| {
+            std::env::var("EDITOR")
+                .ok()
+                .filter(|value| !value.trim().is_empty())
+        })
+        .or_else(|| {
+            fallback_editors()
+                .iter()
+                .copied()
+                .find(|candidate| executable_on_path(candidate))
+                .map(str::to_string)
+        })
+}
+
+fn executable_on_path(name: &str) -> bool {
+    let Some(paths) = std::env::var_os("PATH") else {
+        return false;
+    };
+    std::env::split_paths(&paths).any(|dir| dir.join(name).is_file())
+}
+
+fn fallback_editors() -> &'static [&'static str] {
+    if cfg!(windows) {
+        &["notepad.exe", "nano", "vim"]
+    } else {
+        &["nano", "vi", "vim", "editor"]
+    }
+}
